requirements.txt
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
icrawler==0.6.6
idna==3.3
lxml==4.8.0
Pillow==9.1.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
